third_party/blink/web_tests/svg/custom/invalid-lengthlist-expected.txt

CONSOLE ERROR: Error: <text> attribute x: Expected length, "1.2.3".
